削除内容：jpメソッド集とファイル情報変換メソッド集rubyファイルを削除
理由：可読性向上のためファイル分割したが、コマンド化で全部のファイルが必要で、利用しずらかったため
