+++ 04.ls/lib/my-ls.rb
@@ -4,9 +4,6 @@
 require 'optparse'
 require 'etc'
 
-require_relative 'jp_methods.rb'
-require_relative 'conversion_file_info_methods.rb'
-
 LIST_ROW_NUM = 3
 
 def get_file_names(argument_name, options)
@@ -41,6 +38,19 @@ def select_files(target_dir, target_file, options)
   options['l'] ? get_files_info_text_merged(target_dir, file_names_all) : file_names_all
 end
 
+def sort_jp(jp_array)
+  # Array#sortとは異なり、漢字→ひらがな→カタカナの順にソートするメソッド
+  jp_array.sort do |a, b|
+    if a.match?(/\p{Han}/) && b.match?(/\p{Hiragana}|\p{Katakana}/)
+      -1
+    elsif a.match?(/\p{Hiragana}|\p{Katakana}/) && b.match?(/\p{Han}/)
+      1
+    else
+      a <=> b
+    end
+  end
+end
+
 def get_files_info_text_merged(target_dir, file_names_all)
   files_info_each_type = get_files_info_each_type(target_dir, file_names_all)
   files_info_text = files_info_each_type.values.transpose.map(&:join)
@@ -58,6 +68,75 @@ def get_files_info_each_type(target_dir, file_names_all)
   files_info_each_type['mtime'] = convert_files_mtime_to_l_option_format(files_info.map(&:mtime), 1)
   files_info_each_type['file_name'] = align_jp_str_list_to_left(file_names_all, 0)
   files_info_each_type
+end
+
+def convert_files_mode_to_l_option_format(files_mode, number_of_space)
+  files_mode_chars = files_mode.map do |file_mode|
+    file_mode_bits = format('%016b', file_mode)
+    {
+      'file_type' => convert_file_type_bit_to_char(file_mode_bits[0..3]),
+      'owner_permission' => convert_permission_bits_to_str(file_mode_bits[7..9], file_mode_bits[4]),
+      'group_permission' => convert_permission_bits_to_str(file_mode_bits[10..12], file_mode_bits[5]),
+      'others_permission' => convert_permission_bits_to_str(file_mode_bits[13..15], file_mode_bits[6])
+    }
+  end
+  files_mode_chars.map do |chars|
+    chars['file_type'] + chars['owner_permission'] + chars['group_permission'] + chars['others_permission'] + '@' + ' '*number_of_space
+  end
+end
+
+def convert_file_type_bit_to_char(file_type_bit)
+  file_type_char = { # file_modeに対する、8進数対応表 cf. Linux file type and mode Doc
+    '0o01'.to_i(8) =>	'p', # FIFO
+    '0o02'.to_i(8) => 'c', # Character special file
+    '0o04'.to_i(8) =>	'd', # Directory(ディレクトリ)
+    '0o06'.to_i(8) =>	'b', # Block special file
+    '0o10'.to_i(8) =>	'-', # Regular file(通常ファイル)
+    '0o12'.to_i(8) =>	'l', # Symbolic link(シンボリックリンク)
+    '0o14'.to_i(8) =>	's'	 # Socket link
+  }
+  file_type_char[file_type_bit.to_i(2)]
+end
+
+def convert_permission_bits_to_str(permission_bits, special_permission_bit)
+  file_permission = ['-', '-', '-']
+  file_permission[0] = 'r' if permission_bits[0] == '1'
+  file_permission[1] = 'w' if permission_bits[1] == '1'
+  file_permission[2] = ['-x', 'Ss'][special_permission_bit.to_i][permission_bits[2].to_i] # 2次元テーブルから実行権限記号を選択
+  file_permission.join
+end
+
+def convert_files_mtime_to_l_option_format(files_mtime, number_of_space)
+  files_each_mtime = {}
+  files_each_mtime['month'] = align_str_list_to_right(files_mtime.map(&:month).map(&:to_s), 1)
+  files_each_mtime['day'] = align_str_list_to_right(files_mtime.map(&:day).map(&:to_s), 1)
+  files_each_mtime['time'] = align_str_list_to_right(files_mtime.map { |mtime| "#{mtime.hour}:#{mtime.min}" }, number_of_space)
+  [files_each_mtime['month'], files_each_mtime['day'], files_each_mtime['time']].transpose.map(&:join)
+end
+
+def align_str_list_to_right(str_list, number_of_space)
+  max_size_str = str_list.map(&:size).max
+  str_list.map { |str| str.rjust(max_size_str) + ' '*number_of_space }
+end
+
+def align_jp_str_list_to_left(str_list, number_of_space)
+  max_size_str = str_list.map { |str| size_jp(str) }.max
+  str_list.map { |str| ljust_jp(str, max_size_str) + ' '*number_of_space }
+end
+
+def size_jp(jp_string)
+  # String#sizeと異なり、日本語を2文字とみなすメソッド
+  jp_string.each_char.sum do |char|
+    if char.match?(/\p{Han}|\p{Hiragana}|\p{Katakana}|ー|（|）/)
+      2
+    else
+      1
+    end
+  end
+end
+
+def ljust_jp(jp_string, max_size)
+  jp_string + ' ' * (max_size - size_jp(jp_string))
 end
 
 def generate_name_list_text(file_names, number, options)
